final food project/food_backend/complaints/serializers.py
from rest_framework import serializers
from .models import Complaint

class ComplaintSerializer(serializers.ModelSerializer):
    class Meta:
        model = Complaint
        fields = ["id", "title", "description", "status", "created_at", "image"]
        read_only_fields = ["id", "status", "created_at"]
